requirements.txt
# --- Web Framework & Server ---
fastapi                       # Modern web framework for building the EYAVAP API
uvicorn[standard]             # High-performance ASGI server to run the application

# --- Configuration & Environment ---
python-dotenv                 # Loads secret environment variables from .env
pydantic                      # Data validation using Python type hints
pydantic-settings             # Specialized settings management for Pydantic

# --- Communication (SDK & Networking) ---
requests                      # Standard synchronous HTTP library for quick API calls
httpx                         # Next-generation async HTTP client for agent communication
anyio                         # High-level asynchronous control for concurrent tasks

# --- Security & Authentication ---
python-jose[cryptography]     # Implements JWT tokens to verify AI agent identities
passlib[bcrypt]               # Secure password hashing for agent credentials

# --- Database & Memory (Supabase) ---
supabase                      # Official client for Supabase (The AI memory bank)
postgrest                     # RESTful API client for interacting with the database

# --- Testing & Quality Assurance ---
pytest                        # The standard framework for writing and running system tests
pytest-asyncio                # Enables testing for asynchronous FastAPI endpoints

# --- Logging & Monitoring ---
loguru                        # Advanced, easy-to-use logging for system events
prometheus-client             # Exports system metrics for real-time performance monitoring

# --- Dashboard & Visualization ---
streamlit                     # Interactive web dashboard for monitoring AI agents
pandas                        # Data manipulation and analysis for log processing

# --- Utilities ---
python-dateutil               # Powerful extensions to the standard datetime module
pyyaml                        # Support for reading/writing YAML configuration files
slowapi                       # Rate limiting to protect your API from being overwhelmed